requirements.txt
arrow==1.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.4
pandas==1.4.2
Pillow==9.1.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
six==1.16.0
stravalib==0.10.4
units==0.7
urllib3==1.26.9
